pyproject.toml
[tool.poetry]
name = "feast_sandbox"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8"
pytest = "^6.2.5"
feast = {extras = ["gcp"], version = "^0.14.0"}
SQLAlchemy = "^1.4.25"
PyHive = "^0.6.4"
sasl = "^0.3.1"
thrift-sasl = "^0.4.3"
sqlacodegen = "^2.3.0"
scikit-learn = "^1.0"
parquet-tools = "^0.2.9"
google-cloud-bigquery = {extras = ["bqstorage", "pandas"], version = "^2.28.1"}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
